+++ data_version.go
@@ -1,6 +1,6 @@
 package main
 
-const VERSION = "0.0.1"
+const VERSION = "0.0.2"
 
 func VersionCmd([]string) error {
 	DOut("data version %s\n", VERSION)