+++ dev/changelog.md
@@ -1,6 +1,12 @@
 # data changelog
 
 
+## v0.0.2 2013-11-24
+
+Downloading datasets.
+
+- data get (downloads + installs a dataset)
+
 ## v0.0.1 2013-11-22
 
 Initial version.